+++ Source/BlockGame/Private/Boxes/Stairs.cpp
@@ -1,7 +1,7 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 
-#include "Stairs.h"
+#include "Boxes/Stairs.h"
 
 AStairs::AStairs()
 {
