+++ Source/BlockGame/Private/Weapons/TP_PickUpComponent.cpp
@@ -1,6 +1,6 @@
 // Copyright Epic Games, Inc. All Rights Reserved.
 
-#include "TP_PickUpComponent.h"
+#include "Weapons/TP_PickUpComponent.h"
 
 UTP_PickUpComponent::UTP_PickUpComponent()
 {
