+++ Source/BlockGame/Private/Weapons/TP_WeaponComponent.cpp
@@ -1,14 +1,14 @@
 // Copyright Epic Games, Inc. All Rights Reserved.
 
 
-#include "TP_WeaponComponent.h"
-#include "BlockGameCharacter.h"
-#include "BlockGameProjectile.h"
+#include "Weapons/TP_WeaponComponent.h"
 #include "GameFramework/PlayerController.h"
 #include "Camera/PlayerCameraManager.h"
 #include "Kismet/GameplayStatics.h"
 #include "EnhancedInputComponent.h"
 #include "EnhancedInputSubsystems.h"
+#include "BlockGame/BlockGameCharacter.h"
+#include "Weapons/Projectiles/BlockGameProjectile.h"
 
 // Sets default values for this component's properties
 UTP_WeaponComponent::UTP_WeaponComponent()
