+++ Source/BlockGame/Public/Weapons/Gun.h
@@ -4,7 +4,7 @@
 
 #include "CoreMinimal.h"
 #include "GameFramework/Actor.h"
-#include "RifleProjectile.h"
+#include "Weapons/Projectiles/Projectile.h"
 #include "Components/ArrowComponent.h"
 #include "Gun.generated.h"
 
@@ -39,7 +39,7 @@ protected:
 
 	/** Projectile class to spawn */
 	UPROPERTY(EditDefaultsOnly, Category=Projectile)
-	TSubclassOf<class ARifleProjectile> ProjectileClass;
+	TSubclassOf<class AProjectile> ProjectileClass;
 
 	/** Sound to play each time we fire */
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category=Gameplay)
