+++ Source/BlockGame/Public/Weapons/TP_PickUpComponent.h
@@ -4,7 +4,7 @@
 
 #include "CoreMinimal.h"
 #include "Components/SphereComponent.h"
-#include "BlockGameCharacter.h"
+#include "BlockGame/BlockGameCharacter.h"
 #include "TP_PickUpComponent.generated.h"
 
 // Declaration of the delegate that will be called when someone picks this up
