+++ Source/BlockGame/Public/Weapons/TP_WeaponComponent.h
@@ -4,7 +4,7 @@
 
 #include "CoreMinimal.h"
 #include "Components/SkeletalMeshComponent.h"
-#include "BlockGameProjectile.h"
+#include "Weapons/Projectiles/BlockGameProjectile.h"
 #include "TP_WeaponComponent.generated.h"
 
 class ABlockGameCharacter;
